+++ internal/lido/contracts/contractsAddress.go
@@ -46,8 +46,8 @@ var deployedAddresses = map[string]address{
 		configs.NetworkMainnet: "0xD99CC66fEC647E68294C6477B40fC7E0F6F618D0",
 	},
 	MEVBoostRelayAllowedList: {
-		configs.NetworkMainnet: "0xF95f069F9AD107938F6ba802a3da87892298610E",
 		configs.NetworkHolesky: "0x2d86C5855581194a386941806E38cA119E50aEA3",
+		configs.NetworkMainnet: "0xF95f069F9AD107938F6ba802a3da87892298610E",
 	},
 	Vebo: {
 		configs.NetworkHolesky: "0xffDDF7025410412deaa05E3E1cE68FE53208afcb",